moved remote image code to trait
